+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/PneuService.java
@@ -6,11 +6,8 @@ import br.com.zalf.prolog.webservice.commons.network.Response;
 import br.com.zalf.prolog.webservice.commons.network.ResponseWithCod;
 import br.com.zalf.prolog.webservice.commons.util.Log;
 import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogException;
-import br.com.zalf.prolog.webservice.frota.pneu._model.Pneu;
+import br.com.zalf.prolog.webservice.frota.pneu._model.*;
 import br.com.zalf.prolog.webservice.frota.pneu._model.Pneu.Dimensao;
-import br.com.zalf.prolog.webservice.frota.pneu._model.PneuRetornoDescarte;
-import br.com.zalf.prolog.webservice.frota.pneu._model.PneuRetornoDescarteSuccess;
-import br.com.zalf.prolog.webservice.frota.pneu._model.StatusPneu;
 import br.com.zalf.prolog.webservice.frota.pneu.error.PneuValidator;
 import br.com.zalf.prolog.webservice.frota.pneu.importar.PneuImportReader;
 import br.com.zalf.prolog.webservice.integracao.router.RouterPneu;
@@ -145,15 +142,14 @@ public final class PneuService {
     }
 
     @NotNull
-    public PneuRetornoDescarteSuccess retornarPneuDescarte(@NotNull final PneuRetornoDescarte pneuRetornoDescarte) {
+    public PneuRetornoDescarteResponse retornarPneuDescarte(@NotNull final PneuRetornoDescarte pneuRetornoDescarte) {
         try {
-            return dao.retornarPneuDescarte(pneuRetornoDescarte)
-                    .toBuilder()
-                    .msg("Retorno realizado com sucesso!")
-                    .build();
+            final PneuRetornoDescarteSuccess success = dao.retornarPneuDescarte(pneuRetornoDescarte);
+            return new PneuRetornoDescarteResponse(success.getCodPneuRetornado(),
+                                                   success.getCodMovimentacaoGerada(),
+                                                   "Retorno realizado com sucesso!");
         } catch (final Throwable t) {
-            final String message = String
-                    .format("Erro ao retornar o pneu: %s do descarte.", pneuRetornoDescarte.getCodPneu());
+            final String message = "Erro ao retornar o pneu do descarte.";
             Log.e(TAG, message, t);
             throw Injection
                     .providePneuExceptionHandler()