+++ CHANGELOG.md
@@ -1,18 +1,39 @@
 Change Log
 ==========
 
-### Stage
+<a name="v1.5.3"></a>
+## Version [v1.5.3](https://github.com/luizfp/PrologWebService/compare/v1.5.1...v1.5.3) (release-date) [unreleased]
 
-# Refactor
-
+### Refactors
+* Adapta integração com Visual Rodopar para funcionar na Imediato (PL-3430)
 * Refatorações e reestruturações dos utils
 
-<a name="v1.4.3"></a>
+### Bug Fixes
+* Corrige uso de autoCommit em conexões com o Banco
 
-## Version [v1.4.3](https://github.com/luizfp/PrologWebService/compare/v1.4.2...v1.4.3) (release-date) [unreleased]
+<a name="v1.5.2"></a>
+## Version [v1.5.2](https://github.com/luizfp/PrologWebService/compare/v1.5.1...v1.5.2) (2021-01-21)
+
+### Refactors
+* Melhora forma de resetar uma empresa de apresentação
+
+### Bug Fixes
+* Corrige marshaller de XML na integração com a Praxio (PL-3429)
+
+<a name="v1.5.1"></a>
+## Version [v1.5.1](https://github.com/luizfp/PrologWebService/compare/v1.5.0...v1.5.1) (2021-01-19)
+
+### Refactors
+* Altera forma de lidar com o sentry para utilizar integração com Spring Boot
+* Altera forma de gerar o banner do Prolog no start da aplicação
+
+### Bug Fixes
+* Corrige criação de objetos de response do retorno de pneu de descarte
+
+<a name="v1.5.0"></a>
+## Version [v1.5.0](https://github.com/luizfp/PrologWebService/compare/v1.4.2...v1.5.0) (2021-01-18)
 
 ### Features
-
 * Altera integração da Nepomuceno para trabalhar com código ao invés de placa (PL-3366)
 * Adiciona Spring como gerenciador de conexões, não utilizando mais a classe `DatabaseManager`, configurando através de
   YAML (PL-3265)
@@ -24,14 +45,20 @@ Change Log
 * Corrige busca de histórico de edições (PL-3204)
 
 ### Refactors
+* Adiciona código de veiculo nas movimentações de origem e destino (PL-3321)
 * Adiciona o token_integração em todas as requisições integradas (PL-3251)
 * Refatora objeto de resposta do retorno de pneu de descarte (PL-3371)
 * Adiciona ordenação do relatório de remuneração acumulada com nome do colaborador e cpf (PL-3137)
+* Remove javadocs das DAOs (PL-3400)
+* Remove dependência de pooling do tomcat
+* Refactor altera movimentações para utilizarem código do veículo ao invés da placa (PL-3301)
 
 ### Bug Fixes
 * Relatório qtd aferições by tipo e data não traz primeira linha (PL-3188)
+* Realiza validação mapeamento de exception no internal mapper com sentry (PL-3280)
 
 <a name="v1.4.2"></a>
+
 ## Version [v1.4.2](https://github.com/luizfp/PrologWebService/compare/v1.4.1...v1.4.2) (2020-11-22)
 
 ### Features
@@ -40,6 +67,7 @@ Change Log
 
 ### Bug Fixes
 * Corrige uso de data hora de queries que geram histórico de edição de veículo (PL-3264)
+* Corrige envio de aferição de placas mercosul na integração da Nepomuceno (PL-3392)
 
 ### Refactors
 * Revisa e excluí testes unitários sem sentido (PL-3220)